Split out view specific integration tests to their own test class
